requirements.txt
torch
torchvision
numpy
matplotlib
scikit-image
tqdm
Pillow
torchmetrics